Add files via upload
White noise is a random signal that contains equal intensity across all frequencies within a specified range. In the time domain, it appears as a continuous and random signal where each sample is independent of the others. The name "white noise" is derived from the analogy with white light in optics, where white light contains all visible colors in equal proportions.